+++ tests/test_utils/events/diff_utils/manageDiff.test.js
@@ -26,8 +26,8 @@ describe("manageDiff", () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepo, { recursive: true });
-        fs.rmdirSync(repoPath, { recursive: true });
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepo, { recursive: true, force: true });
     });
 
     test("should be skipped",() => {