+++ tests/test_utils/events/utils/handleNewFile.test.js
@@ -40,8 +40,8 @@ describe("handleNewFile",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepo, { recursive: true });
-        fs.rmdirSync(repoPath, { recursive: true });
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepo, { recursive: true, force: true });
     });
 
     test("Valid File",  async () => {