+++ tests/test_utils/events/utils/handleRename.test.js
@@ -51,8 +51,8 @@ describe("handleNewFile",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepo, { recursive: true });
-        fs.rmdirSync(repoPath, { recursive: true });
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepo, { recursive: true, force: true });
     });
 
     test("handleRename for File",  () => {