+++ tests/test_utils/events/utils/repoIsNotSynced.test.js
@@ -21,8 +21,8 @@ describe("repoIsNotSynced", () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepo, {recursive: true});
-        fs.rmdirSync(repoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepo, { recursive: true, force: true });
     });
 
     test("with no config.yml", () => {