+++ tests/test_utils/events/utils/shouldIgnoreFile.test.js
@@ -20,8 +20,8 @@ beforeAll(() => {
 });
 
 afterAll(() => {
-    fs.rmdirSync(baseRepo, { recursive: true });
-    fs.rmdirSync(repoPath, { recursive: true });
+    fs.rmSync(repoPath, { recursive: true, force: true });
+    fs.rmSync(baseRepo, { recursive: true, force: true });
 });
 
 test("shouldIgnoreFile with git file",  () => {