+++ tests/test_utils/init/utils.test.js
@@ -81,8 +81,8 @@ describe("successfullySynced",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, {recursive: true});
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("Non-Synced Repo",  () => {
@@ -140,8 +140,8 @@ describe("getSyncablePaths",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, {recursive: true});
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("No .syncignore",  () => {
@@ -198,8 +198,8 @@ describe("copyFilesTo",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, {recursive: true});
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("Copy to Shadow repo",  () => {
@@ -232,7 +232,7 @@ describe("saveIamUser",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("With no user.yml",  () => {
@@ -270,7 +270,7 @@ describe("saveSequenceTokenFile",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("With no sequence_token.yml",  () => {
@@ -307,8 +307,8 @@ describe("saveFileIds",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, {recursive: true});
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("Should save file IDs",  () => {
@@ -342,8 +342,8 @@ describe("uploadRepo",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, {recursive: true});
-        fs.rmdirSync(baseRepoPath, {recursive: true});
+        fs.rmSync(repoPath, { recursive: true, force: true });
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("Server Down",  async () => {