+++ tests/test_utils/utils/api_utils.test.js
@@ -22,7 +22,7 @@ describe('checkServerDown', () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(baseRepoPath, { recursive: true});
+        fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("with status: true", async () => {