+++ tests/test_utils/utils/common/getSkipRepos.test.js
@@ -18,7 +18,7 @@ describe("getSkipRepos",  () => {
     });
 
     afterEach(() => {
-        fs.rmdirSync(repoPath, { recursive: true });
+        fs.rmSync(repoPath, { recursive: true, force: true });
     });
 
     test('skipRepos with .syncignore items', () => {