+++ tests/test_utils/utils/common/getSyncIgnoreItems.test.js
@@ -11,7 +11,7 @@ beforeAll(() => {
 });
 
 afterAll(() => {
-    fs.rmdirSync(repoPath, { recursive: true });
+    fs.rmSync(repoPath, { recursive: true, force: true });
 });
 
 test('syncIgnore items with .syncignore', () => {