+++ tests/test_utils/utils/common/readFile.test.js
@@ -12,7 +12,7 @@ beforeEach(() => {
 });
 
 afterEach(() => {
-    fs.rmdirSync(repoPath, { recursive: true });
+    fs.rmSync(repoPath, { recursive: true, force: true });
 });
 
 test('reads file', () => {